sdms-main/backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
SQLAlchemy==2.0.31
python-dotenv==1.0.1
gunicorn==22.0.0
firebase-admin>=6.5.0
PyJWT>=2.8.0
